+++ init/init_descriptions/waxwell.lua
@@ -8,3 +8,5 @@ DESCRIBE = GLOBAL.STRINGS.CHARACTERS.WAXWELL.DESCRIBE
 		{
 			"I need to bring this heart to their rotting corpse.",
 		}
+
+GLOBAL.STRINGS.CHARACTERS.WAXWELL.ANNOUNCE_WINONAGEN = "I'm more of a magic and dark arts type of guy"