+++ init/init_prefabpostinit.lua
@@ -14,7 +14,12 @@ sculpture_???
 ]]--
 
 --TODO: Make lava give more damage when on top
---TODO: Make ponds passable as well, and they make you wet really fast
+
+AddPrefabPostInit("endtable", function(inst)
+    if inst~= nil and inst.Physics ~= nil then
+        GLOBAL.RemovePhysicsColliders(inst)
+    end
+end)
 
 AddPrefabPostInit("fossil_stalker", function(inst)
     if inst~= nil and inst.Physics ~= nil then